Extract generic TabBase from QueryTabEditorTab
