+++ src/panels/QueryEditorTab.ts
@@ -5,206 +5,15 @@
 
 import { callWithTelemetryAndErrorHandling } from '@microsoft/vscode-azext-utils';
 import * as crypto from 'crypto';
-import * as path from 'path';
-import { v4 as uuid } from 'uuid';
 import * as vscode from 'vscode';
 import { getNoSqlQueryConnection } from '../docdb/commands/connectNoSqlContainer';
-import { type NoSqlQueryConnection } from '../docdb/NoSqlCodeLensProvider';
 import { CosmosDBSession } from '../docdb/session/CosmosDBSession';
 import { type ResultViewMetadata } from '../docdb/types/queryResult';
-import { ext } from '../extensionVariables';
-import { TelemetryContext } from '../Telemetry';
 import * as vscodeUtil from '../utils/vscodeUtils';
-import { type Channel } from './Communication/Channel/Channel';
-import { VSCodeChannel } from './Communication/Channel/VSCodeChannel';
+import { TabBase, type CommandPayload } from './TabBase';
 
-const DEV_SERVER_HOST = 'http://localhost:18080';
-
-type CommandPayload = {
-    commandName: string;
-    params: unknown[];
-};
-
-export class QueryEditorTab {
-    public static readonly title = 'Query Editor';
-    public static readonly viewType = 'cosmosDbQuery';
-    public static readonly openTabs: Set<QueryEditorTab> = new Set<QueryEditorTab>();
-
-    public readonly channel: Channel;
-    public readonly panel: vscode.WebviewPanel;
-    public readonly sessions = new Map<string, CosmosDBSession>();
-
-    private readonly id: string;
-    private readonly start: number;
-    private readonly telemetryContext: TelemetryContext;
-
-    private connection: NoSqlQueryConnection | undefined;
-    private disposables: vscode.Disposable[] = [];
-
-    private constructor(panel: vscode.WebviewPanel, connection?: NoSqlQueryConnection) {
-        QueryEditorTab.openTabs.add(this);
-
-        this.id = uuid();
-        this.start = Date.now();
-        this.telemetryContext = new TelemetryContext(connection);
-
-        this.channel = new VSCodeChannel(panel.webview);
-        this.panel = panel;
-        this.connection = connection;
-
-        this.panel.onDidDispose(() => this.dispose(), null, this.disposables);
-
-        this.panel.webview.html = this.getWebviewContent();
-
-        this.initController();
-
-        void this.telemetryContext.reportWebviewEvent('webviewOpened', {
-            panelId: this.id,
-            hasConnection: connection ? 'true' : 'false',
-        });
-    }
-
-    public static render(connection?: NoSqlQueryConnection, viewColumn?: vscode.ViewColumn): QueryEditorTab {
-        const column = viewColumn ?? vscode.ViewColumn.One;
-        if (connection) {
-            const openTab = [...QueryEditorTab.openTabs].find(
-                (openTab) =>
-                    openTab.connection?.endpoint === connection.endpoint &&
-                    openTab.connection?.databaseId === connection.databaseId &&
-                    openTab.connection?.containerId === connection.containerId,
-            );
-            if (openTab) {
-                openTab.panel.reveal(column);
-                return openTab;
-            }
-        }
-
-        const panel = vscode.window.createWebviewPanel(QueryEditorTab.viewType, QueryEditorTab.title, column, {
-            enableScripts: true,
-            retainContextWhenHidden: true,
-        });
-
-        return new QueryEditorTab(panel, connection);
-    }
-
-    public dispose(): void {
-        QueryEditorTab.openTabs.delete(this);
-
-        this.channel.dispose();
-        this.panel.dispose();
-
-        this.sessions.forEach((session) => session.dispose());
-        this.sessions.clear();
-
-        while (this.disposables.length) {
-            const disposable = this.disposables.pop();
-            if (disposable) {
-                disposable.dispose();
-            }
-        }
-
-        void this.telemetryContext.reportWebviewEvent(
-            'webviewClosed',
-            {
-                panelId: this.id,
-            },
-            { openedTime: (Date.now() - this.start) / 1000 },
-        );
-    }
-
-    private getWebviewContent(): string {
-        const ctx = ext.context;
-        const cspSource = this.panel.webview.cspSource;
-        const devServer = !!process.env.DEVSERVER;
-        const isProduction = ext.context.extensionMode === vscode.ExtensionMode.Production;
-        const nonce = crypto.randomBytes(16).toString('base64');
-
-        const dir = ext.isBundle ? '' : 'out/src/webviews';
-        const filename = ext.isBundle ? 'views.js' : 'index.js';
-        const uri = (...parts: string[]) =>
-            this.panel.webview
-                .asWebviewUri(vscode.Uri.file(path.join(ctx.extensionPath, dir, ...parts)))
-                .toString(true);
-
-        const publicPath = isProduction || !devServer ? uri() : `${DEV_SERVER_HOST}/`;
-        const srcUri = isProduction || !devServer ? uri(filename) : `${DEV_SERVER_HOST}/${filename}`;
-
-        const csp = (
-            isProduction
-                ? [
-                      `form-action 'none';`,
-                      `default-src ${cspSource};`,
-                      `script-src ${cspSource} 'nonce-${nonce}';`,
-                      `style-src ${cspSource} ${DEV_SERVER_HOST} 'unsafe-inline';`,
-                      `font-src ${cspSource} ${DEV_SERVER_HOST};`,
-                      `worker-src ${cspSource} ${DEV_SERVER_HOST} blob:;`,
-                      `img-src ${cspSource} ${DEV_SERVER_HOST} data:;`,
-                  ]
-                : [
-                      `form-action 'none';`,
-                      `default-src ${cspSource} ${DEV_SERVER_HOST};`,
-                      `style-src ${cspSource} ${DEV_SERVER_HOST} 'unsafe-inline';`,
-                      `script-src ${cspSource} ${DEV_SERVER_HOST} 'nonce-${nonce}';`,
-                      `connect-src ${cspSource} ${DEV_SERVER_HOST} ws:;`,
-                      `font-src ${cspSource} ${DEV_SERVER_HOST};`,
-                      `worker-src ${cspSource} ${DEV_SERVER_HOST} blob:;`,
-                      `img-src ${cspSource} ${DEV_SERVER_HOST} data:;`,
-                  ]
-        ).join(' ');
-
-        return this.template({
-            title: QueryEditorTab.title,
-            csp,
-            srcUri,
-            publicPath,
-            viewType: QueryEditorTab.viewType,
-            nonce,
-        });
-    }
-
-    private template(params: {
-        csp: string;
-        viewType: string;
-        srcUri: string;
-        publicPath: string;
-        title: string;
-        nonce: string;
-    }) {
-        return `
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>${params.title}</title>
-    <meta http-equiv="Content-Security-Policy" content="${params.csp}" />
-  </head>
-
-  <body>
-    <div id="root"></div>
-    <script type="module" nonce="${params.nonce}">
-      import { render } from "${params.srcUri}";
-      render("${params.viewType}", acquireVsCodeApi(), "${params.publicPath}");
-    </script>
-  </body>
-</html>
-`;
-    }
-
-    private initController() {
-        this.channel.on<void>('command', async (payload: CommandPayload) => {
-            // TODO: Telemetry
-            console.log('command', payload);
-
-            await this.getCommand(payload);
-        });
-
-        this.channel.on<void>('ready', async () => {
-            await this.updateConnection(this.connection);
-        });
-    }
-
-    private getCommand(payload: CommandPayload): Promise<void> {
+export class QueryEditorTab extends TabBase {
+    protected getCommand(payload: CommandPayload): Promise<void> {
         const commandName = payload.commandName;
         switch (commandName) {
             case 'openFile':
@@ -252,30 +61,6 @@ export class QueryEditorTab {
                 return (async () => await vscode.commands.executeCommand('azureDatabases.reportIssue'))();
             default:
                 throw new Error(`Unknown command: ${commandName}`);
-        }
-    }
-
-    private async updateConnection(connection?: NoSqlQueryConnection): Promise<void> {
-        this.connection = connection;
-
-        if (this.connection) {
-            const { databaseId, containerId, endpoint, masterKey } = this.connection;
-
-            this.telemetryContext.addMaskedValue([databaseId, containerId, endpoint, masterKey ?? '']);
-
-            await this.channel.postMessage({
-                type: 'event',
-                name: 'databaseConnected',
-                params: [databaseId, containerId],
-            });
-        } else {
-            // We will not remove the connection details from the telemetry context
-            // to prevent accidental logging of sensitive information
-            await this.channel.postMessage({
-                type: 'event',
-                name: 'databaseDisconnected',
-                params: [],
-            });
         }
     }
